databases/requirements.txt
pymongo
psycopg2
mysql-connector-python
pyodbc